NFC: Extract BlockCreator from MPC
