+++ casper/src/main/scala/coop/rchain/casper/BlockCreator.scala
@@ -1,0 +1,210 @@
+package coop.rchain.casper
+
+import cats.Monad
+import cats.effect.Sync
+import cats.implicits._
+import coop.rchain.blockstorage.{BlockDagRepresentation, BlockStore}
+import coop.rchain.casper.protocol._
+import coop.rchain.casper.util.{DagOperations, ProtoUtil}
+import coop.rchain.casper.util.ProtoUtil.{
+  blockHeader,
+  bonds,
+  chooseNonConflicting,
+  signBlock,
+  toJustification,
+  unsignedBlockProto
+}
+import coop.rchain.casper.util.rholang.RuntimeManager.StateHash
+import coop.rchain.casper.util.rholang._
+import coop.rchain.shared.{Cell, Log, Time}
+
+object BlockCreator {
+  def createBlock[F[_]: Sync: Log: Time: BlockStore](
+      dag: BlockDagRepresentation[F],
+      genesis: BlockMessage,
+      publicKey: Array[Byte],
+      privateKey: Array[Byte],
+      sigAlgorithm: String,
+      shardId: String,
+      version: Long,
+      expirationThreshold: Int,
+      runtimeManager: RuntimeManager[F]
+  )(implicit state: Cell[F, CasperState]): F[CreateBlockStatus] = {
+    def updateDeployValidAfterBlock(deployData: DeployData, max: Long): DeployData =
+      if (deployData.validAfterBlockNumber == -1)
+        deployData.withValidAfterBlockNumber(max)
+      else
+        deployData
+
+    def updateDeployHistory(state: CasperState, max: Long): CasperState =
+      state.copy(deployHistory = state.deployHistory.map(deployData => {
+        updateDeployValidAfterBlock(deployData, max)
+      }))
+
+    for {
+      tipHashes <- Estimator.tips[F](dag, genesis)
+      p         <- chooseNonConflicting[F](tipHashes, dag)
+      _ <- Log[F].info(
+            s"${p.size} parents out of ${tipHashes.size} latest blocks will be used."
+          )
+      maxBlockNumber = ProtoUtil.maxBlockNumber(p)
+      /*
+       * This mechanism is a first effort to make life of a deploying party easier.
+       * Instead of expecting the user to guess the current block number we assume that
+       * if no value is given (default: -1) rchain should try to deploy
+       * with the current known max block number.
+       *
+       * TODO make more developer friendly by introducing Option instead of a magic number
+       */
+      _                <- state.modify(state => updateDeployHistory(state, maxBlockNumber))
+      r                <- remDeploys(dag, p, maxBlockNumber, expirationThreshold)
+      bondedValidators = bonds(p.head).map(_.validator).toSet
+      //We ensure that only the justifications given in the block are those
+      //which are bonded validators in the chosen parent. This is safe because
+      //any latest message not from a bonded validator will not change the
+      //final fork-choice.
+      latestMessages <- dag.latestMessages
+      justifications = toJustification(latestMessages)
+        .filter(j => bondedValidators.contains(j.validator))
+      proposal <- if (r.nonEmpty || p.length > 1) {
+                   createProposal[F](
+                     dag,
+                     runtimeManager,
+                     p,
+                     r,
+                     justifications,
+                     maxBlockNumber,
+                     shardId,
+                     version
+                   )
+                 } else {
+                   CreateBlockStatus.noNewDeploys.pure[F]
+                 }
+      signedBlock <- proposal.mapF(
+                      signBlock(_, dag, publicKey, privateKey, sigAlgorithm, shardId)
+                    )
+    } yield signedBlock
+  }
+
+  // TODO: Remove no longer valid deploys here instead of with lastFinalizedBlock call
+  private def remDeploys[F[_]: Monad: Log: Time: BlockStore](
+      dag: BlockDagRepresentation[F],
+      parents: Seq[BlockMessage],
+      maxBlockNumber: Long,
+      expirationThreshold: Int
+  )(implicit state: Cell[F, CasperState]): F[Seq[DeployData]] =
+    for {
+      state               <- state.read
+      currentBlockNumber  = maxBlockNumber + 1
+      earliestBlockNumber = currentBlockNumber - expirationThreshold
+      deploys             = state.deployHistory
+      validDeploys = deploys.filter(
+        d => notFutureDeploy(currentBlockNumber, d) && notExpiredDeploy(earliestBlockNumber, d)
+      )
+      deploysInCurrentChain <- DagOperations
+                                .bfTraverseF[F, BlockMessage](parents.toList)(
+                                  b =>
+                                    ProtoUtil
+                                      .unsafeGetParentsAboveBlockNumber[F](b, earliestBlockNumber)
+                                )
+                                .map { b =>
+                                  ProtoUtil.deploys(b).flatMap(_.deploy)
+                                }
+                                .toList
+    } yield (validDeploys -- deploysInCurrentChain.flatten).toSeq
+
+  private def notExpiredDeploy(earliestBlockNumber: Long, d: DeployData): Boolean =
+    d.validAfterBlockNumber > earliestBlockNumber
+
+  private def notFutureDeploy(currentBlockNumber: Long, d: DeployData): Boolean =
+    d.validAfterBlockNumber < currentBlockNumber
+
+  private def createProposal[F[_]: Sync: Log: Time: BlockStore](
+      dag: BlockDagRepresentation[F],
+      runtimeManager: RuntimeManager[F],
+      p: Seq[BlockMessage],
+      r: Seq[DeployData],
+      justifications: Seq[Justification],
+      maxBlockNumber: Long,
+      shardId: String,
+      version: Long
+  ): F[CreateBlockStatus] =
+    for {
+      now <- Time[F].currentMillis
+      possibleProcessedDeploys <- InterpreterUtil.computeDeploysCheckpoint[F](
+                                   p,
+                                   r,
+                                   dag,
+                                   runtimeManager,
+                                   Some(now)
+                                 )
+      result <- possibleProcessedDeploys match {
+                 case Left(ex) =>
+                   Log[F]
+                     .error(
+                       s"Critical error encountered while processing deploys: ${ex.getMessage}"
+                     )
+                     .map(_ => CreateBlockStatus.internalDeployError(ex))
+
+                 case Right((preStateHash, postStateHash, processedDeploys)) =>
+                   val (internalErrors, persistableDeploys) =
+                     processedDeploys.partition(_.status.isInternalError)
+                   internalErrors.toList
+                     .traverse {
+                       case InternalProcessedDeploy(deploy, _, _, InternalErrors(errors)) =>
+                         val errorsMessage = errors.map(_.getMessage).mkString("\n")
+                         Log[F].error(
+                           s"Internal error encountered while processing deploy ${PrettyPrinter
+                             .buildString(deploy)}: $errorsMessage"
+                         )
+                       case _ => ().pure[F]
+                     }
+                     .flatMap(_ => {
+                       runtimeManager
+                         .computeBonds(postStateHash)
+                         .map { newBonds =>
+                           createBlock(
+                             now,
+                             p,
+                             justifications,
+                             maxBlockNumber,
+                             preStateHash,
+                             postStateHash,
+                             persistableDeploys,
+                             newBonds,
+                             shardId,
+                             version
+                           )
+                         }
+                     })
+               }
+    } yield result
+
+  private def createBlock(
+      now: Long,
+      p: Seq[BlockMessage],
+      justifications: Seq[Justification],
+      maxBlockNumber: Long,
+      preStateHash: StateHash,
+      postStateHash: StateHash,
+      persistableDeploys: Seq[InternalProcessedDeploy],
+      newBonds: Seq[Bond],
+      shardId: String,
+      version: Long
+  ): CreateBlockStatus = {
+    val postState = RChainState()
+      .withPreStateHash(preStateHash)
+      .withPostStateHash(postStateHash)
+      .withBonds(newBonds)
+      .withBlockNumber(maxBlockNumber + 1)
+
+    val body = Body()
+      .withState(postState)
+      .withDeploys(
+        persistableDeploys.map(ProcessedDeployUtil.fromInternal)
+      )
+    val header = blockHeader(body, p.map(_.blockHash), version, now)
+    val block  = unsignedBlockProto(body, header, justifications, shardId)
+    CreateBlockStatus.created(block)
+  }
+}
