+++ casper/src/main/scala/coop/rchain/casper/MultiParentCasperImpl.scala
@@ -176,52 +176,19 @@ class MultiParentCasperImpl[F[_]: Sync: Concurrent: Sync: ConnectionsCell: Trans
    */
   def createBlock: F[CreateBlockStatus] = validatorId match {
     case Some(ValidatorIdentity(publicKey, privateKey, sigAlgorithm)) =>
-      def updateDeployValidAfterBlock(deployData: DeployData, max: Long) =
-        if (deployData.validAfterBlockNumber == -1)
-          deployData.withValidAfterBlockNumber(max)
-        else
-          deployData
-
-      def updateDeployHistory(state: CasperState, max: Long) =
-        state.copy(deployHistory = state.deployHistory.map(deployData => {
-          updateDeployValidAfterBlock(deployData, max)
-        }))
-
-      for {
-        dag       <- blockDag
-        tipHashes <- estimator(dag)
-        p         <- chooseNonConflicting[F](tipHashes, dag)
-        _ <- Log[F].info(
-              s"${p.size} parents out of ${tipHashes.size} latest blocks will be used."
-            )
-        maxBlockNumber = ProtoUtil.maxBlockNumber(p)
-        /*
-         * This mechanism is a first effort to make life of a deploying party easier.
-         * Instead of expecting the user to guess the current block number we assume that
-         * if no value is given (default: -1) rchain should try to deploy
-         * with the current known max block number.
-         *
-         * TODO make more developer friendly by introducing Option instead of a magic number
-         */
-        _                <- Cell[F, CasperState].modify(state => updateDeployHistory(state, maxBlockNumber))
-        r                <- remDeploys(dag, p, maxBlockNumber)
-        bondedValidators = bonds(p.head).map(_.validator).toSet
-        //We ensure that only the justifications given in the block are those
-        //which are bonded validators in the chosen parent. This is safe because
-        //any latest message not from a bonded validator will not change the
-        //final fork-choice.
-        latestMessages <- dag.latestMessages
-        justifications = toJustification(latestMessages)
-          .filter(j => bondedValidators.contains(j.validator))
-        proposal <- if (r.nonEmpty || p.length > 1) {
-                     createProposal(dag, p, r, justifications, maxBlockNumber)
-                   } else {
-                     CreateBlockStatus.noNewDeploys.pure[F]
-                   }
-        signedBlock <- proposal.mapF(
-                        signBlock(_, dag, publicKey, privateKey, sigAlgorithm, shardId)
-                      )
-      } yield signedBlock
+      BlockDagStorage[F].getRepresentation.flatMap { dag =>
+        BlockCreator.createBlock(
+          dag,
+          genesis,
+          publicKey,
+          privateKey,
+          sigAlgorithm,
+          shardId,
+          version,
+          expirationThreshold,
+          runtimeManager
+        )
+      }
     case None => CreateBlockStatus.readOnlyMode.pure[F]
   }
 
@@ -234,120 +201,6 @@ class MultiParentCasperImpl[F[_]: Sync: Concurrent: Sync: ConnectionsCell: Trans
       _            <- lastFinalizedBlockHashContainer.set(updatedLastFinalizedBlockHash)
       blockMessage <- ProtoUtil.unsafeGetBlock[F](updatedLastFinalizedBlockHash)
     } yield blockMessage
-
-  // TODO: Remove no longer valid deploys here instead of with lastFinalizedBlock call
-  private def remDeploys(
-      dag: BlockDagRepresentation[F],
-      parents: Seq[BlockMessage],
-      maxBlockNumber: Long
-  ): F[Seq[DeployData]] =
-    for {
-      state               <- Cell[F, CasperState].read
-      currentBlockNumber  = maxBlockNumber + 1
-      earliestBlockNumber = currentBlockNumber - expirationThreshold
-      deploys             = state.deployHistory
-      validDeploys = deploys.filter(
-        d => notFutureDeploy(currentBlockNumber, d) && notExpiredDeploy(earliestBlockNumber, d)
-      )
-      deploysInCurrentChain <- DagOperations
-                                .bfTraverseF[F, BlockMessage](parents.toList)(
-                                  b =>
-                                    ProtoUtil
-                                      .unsafeGetParentsAboveBlockNumber[F](b, earliestBlockNumber)
-                                )
-                                .map { b =>
-                                  ProtoUtil.deploys(b).flatMap(_.deploy)
-                                }
-                                .toList
-    } yield (validDeploys -- deploysInCurrentChain.flatten).toSeq
-
-  private def notExpiredDeploy(earliestBlockNumber: Long, d: DeployData): Boolean =
-    d.validAfterBlockNumber > earliestBlockNumber
-
-  private def notFutureDeploy(currentBlockNumber: Long, d: DeployData): Boolean =
-    d.validAfterBlockNumber < currentBlockNumber
-
-  private def createProposal(
-      dag: BlockDagRepresentation[F],
-      p: Seq[BlockMessage],
-      r: Seq[DeployData],
-      justifications: Seq[Justification],
-      maxBlockNumber: Long
-  ): F[CreateBlockStatus] =
-    for {
-      now <- Time[F].currentMillis
-      possibleProcessedDeploys <- InterpreterUtil.computeDeploysCheckpoint[F](
-                                   p,
-                                   r,
-                                   dag,
-                                   runtimeManager,
-                                   Some(now)
-                                 )
-      result <- possibleProcessedDeploys match {
-                 case Left(ex) =>
-                   Log[F]
-                     .error(
-                       s"Critical error encountered while processing deploys: ${ex.getMessage}"
-                     )
-                     .map(_ => CreateBlockStatus.internalDeployError(ex))
-
-                 case Right((preStateHash, postStateHash, processedDeploys)) =>
-                   val (internalErrors, persistableDeploys) =
-                     processedDeploys.partition(_.status.isInternalError)
-                   internalErrors.toList
-                     .traverse {
-                       case InternalProcessedDeploy(deploy, _, _, InternalErrors(errors)) =>
-                         val errorsMessage = errors.map(_.getMessage).mkString("\n")
-                         Log[F].error(
-                           s"Internal error encountered while processing deploy ${PrettyPrinter
-                             .buildString(deploy)}: $errorsMessage"
-                         )
-                       case _ => ().pure[F]
-                     }
-                     .flatMap(_ => {
-                       runtimeManager
-                         .computeBonds(postStateHash)
-                         .map { newBonds =>
-                           createBlock(
-                             now,
-                             p,
-                             justifications,
-                             maxBlockNumber,
-                             preStateHash,
-                             postStateHash,
-                             persistableDeploys,
-                             newBonds
-                           )
-                         }
-                     })
-               }
-    } yield result
-
-  private def createBlock(
-      now: Long,
-      p: Seq[BlockMessage],
-      justifications: Seq[Justification],
-      maxBlockNumber: Long,
-      preStateHash: StateHash,
-      postStateHash: StateHash,
-      persistableDeploys: Seq[InternalProcessedDeploy],
-      newBonds: Seq[Bond]
-  ): CreateBlockStatus = {
-    val postState = RChainState()
-      .withPreStateHash(preStateHash)
-      .withPostStateHash(postStateHash)
-      .withBonds(newBonds)
-      .withBlockNumber(maxBlockNumber + 1)
-
-    val body = Body()
-      .withState(postState)
-      .withDeploys(
-        persistableDeploys.map(ProcessedDeployUtil.fromInternal)
-      )
-    val header = blockHeader(body, p.map(_.blockHash), version, now)
-    val block  = unsignedBlockProto(body, header, justifications, shardId)
-    CreateBlockStatus.created(block)
-  }
 
   def blockDag: F[BlockDagRepresentation[F]] =
     BlockDagStorage[F].getRepresentation
